Server: make job handle safer
Use a global counter in the job handle.  Previously the job handle
was using the hostname + a monotonically increasing counter for
the connection.  If there were two connections from the same host,
duplicate handles would be generated.  Instead, use a global counter
for the job handle to guarantee that no collisions will happen.

The hostname portion of the handle is no longer needed, but is
retained for format compatibility with C gearman (in case some
tool assumes something about the format).

Change-Id: I44d1784b97d78fc1879a5612316165fa49b90b83