Fix geard logging config
When passing in a logging config file geard was throwing the
following error: 'module' object has no attribute 'config'
The fix is to import 'logging.config'

Logs were not being written to log resources when running
in daemon mode because the dameon closed all of the
previously opened log resources.  The fix is to setup
logging after running the daemon.

Change-Id: Ie3e4139e26712a36894d16b48bfe3bca6d5778a9