Solve hacking/pbr requirements conflict
* requirements.txt: Add an empty line making it more clear that pbr
has to appear first in the requirements list, before other entries.
Also specify a newer required minimum pbr since the old one had
bugs and ought to be avoided.

* setup.py: Stop declaring the pbr version range here, since it's
handled separately by the requirements list.

* test-requirements.txt: Allow newer hacking with solvable pbr
second-order requirement and move it to the top of the list so that
it gets satisfied before other test requirements.

Change-Id: I4f80094c719f43a1261091263ccf60eefe5d9af1