Use client_id as part of the logger name
In a system with many gearman clients in a single process, it is
helpful to designate them with client_ids which show up in the
logs (otherwise it can be impossible to determine, say, which
gearman client is waiting indefinitely for a packet).  Add them
to the logger name so that admins can route log messages based
on client ids.

Change-Id: Ia51762ca063d910f8771209a63a50bc418d04004