Restructure logging
Rework logging so that an operator can get a fairly complete picture
of events at INFO level.  All packets in and out are logged at INFO,
as well as connects and disconnects.

When logging packets, log some interesting arguments with them as well
to aid in matching traffic up with other systems.

Don't log the exception that goes along with "Connection reset by
peer" when reading from the network (just log that the connection
was closed).

A server instance usually has no "client_id" associated with it, so
default to not using the append-client_id-to-logger-name feature
with the server.

Change-Id: I3f1198cc5e9d80077f5f86a663c646c88ba87f99