Support binding to a specified host name or address
As it happens, my CentOS7 system has broken getaddrinfo which apparently
prefers IPv4-only sockets:

 >>> socket.getaddrinfo(None, 12345, socket.AF_UNSPEC,
                        socket.SOCK_STREAM, 0, socket.AI_PASSIVE)
 [(2, 1, 6, '', ('0.0.0.0', 12345)),
  (10, 1, 6, '', ('::', 12345, 0, 0))]

That's kind of a bummer on a host which is supposed to be reachable only
through IPv6.

I don't feel like waiting for
https://sourceware.org/bugzilla/show_bug.cgi?id=9981 to get fixed, and
binding to a correct socket sounds like a good idea anyway, hence this
patch. I'll be extending zuul with the same feature as well.

Change-Id: I884e0ddcaac9c7de35b913dc967a47ee854d0a8a