Fix accounting bug when peeking for job
When peeking to see if a job is available, we erroneously would
mark the job as running.  This likely resulting in a small bug
in job accounting as one more job than was actually running for
a given function might be reported.  Also, statsd data would have
been emitted when not strictly necessary.

Change-Id: I741a1a1bebef822d42ad0f27342d40b37b5e1778