Fix SSL non-blocking IO support
When an SSL socket is in non-blocking mode, EAGAIN is essentially
masked by WANT_READ or WANT_WRITE.  So treat EAGAIN, WANT_READ,
and WANT_WRITE as equivalent.  The reason for treating them as
equivalent in non-blocking mode is that the following cases can
happen:

While reading, we might get:
  SSL_WANT_READ:
    This is the same as EAGAIN.
  SSL_WANT_WRITE:
    We need to wait until the edge poll triggers with POLLOUT
    signalling that we can write again.  When that happens,
    we should retry this read call.

While writing, we might get:
  SSL_WANT_READ:
    We need to wait until the edge poll triggers with POLLIN
    signalling that we can read again.  When that happens,
    we should retry this write call.
  SSL_WANT_WRITE:
    This is the same as EAGAIN.

This suggests that any time we receive either of the SSL_WANT
responses, we should wait until either IN or OUT is edge-triggered,
and always attempt to both read and write every time IN or OUT
is triggered.

To do this, mask WANT_READ, WANT_WRITE, and EAGAIN with a new
exception that simply indicates internally to gear to retry the
IO operation after the next poll trigger.

Change-Id: Ib2be2fc97f13a95b4d082a54222013be4ecd4384