Pin python-daemon to <2.0
The 2.0 release series of python-daemon has been broken for a few
reasons. At one point it had an undeclared dependency on docutils
that caused things to break. Current issue is it breaks the egg_info
writers of other packages like extras and pyyaml.

Just pin to a known working release for now so that we can fix
nodepool's lack of timeouts.

Change-Id: I01dbdde344a9fd4086d4013be087fa2b44c2d0a7
Co-Authored-By: Clark Boylan <clark.boylan@gmail.com>