Move running check in getJob
The previous change correctly fixed a race condition, however, if
that code path is used, then the accounting in self.waiting_for_jobs
could be incorrect.  That's not very important because at the moment
restarting a worker is not supported, so that variable will not be
used again.  However, if we move the check a little earlier, we can
keep the accounting accurate while still handling the race condition.

Change-Id: Ide6f1cd16f72ce7e92dbf2f78e370195afb8a334