Do not raise RetryIOError on blocked write
When EAGAIN is returned on a socket write, do not raise RetryIOError.
This exception is used principally to indicate that we have encountered
a blocked read, and a method farther up the call stack may need to
perform some cleanup.  However, within the sendQueuedData method,
all necessary cleanup is performed by the finally handler itself.
Therefore, it is safe to ignore a blocked write, and count on either
a subsequent write (which will append data to the queue and retry
sending the oldest data) or the poll edge trigger to indicate that
we should retry the write.

There is nothing in the call stack above this method that has an
exception handler for RetryIOError.  Therefore, in the current state,
the raised exception was causing connections to be disconnected
due to the error.

Change-Id: I9211fb6365f8f3b6dd0310430cf97926ce1f5d07