Add a send lock to the base Connection class
The sendRaw method (and therefore sendPacket) was originally
thread safe by virtue of consisting of a single socket.send()
call, but when we added SSL, we added a loop within the method
to handle the increased likelihood that not all data would be
sent in one call.  Of course, the method should have been written
this way to start with.

However, this means that we can end up with partial packets being
sent before a context switch to another thread which may also
want to send a packet.  To handle that case, place the entire
method in a lock.

Note, this doesn't affect server connections as they use a
non-blocking connection which has a send queue, so only one thread
ever actuall transmits.

Change-Id: I3bda6fda5f762d18f28b56a43b7dc28f37dbc427